Move Migrations to Migration.
